+++ src/HearThis/UI/UnitNavigationButton.cs
@@ -1,6 +1,6 @@
 // --------------------------------------------------------------------------------------------
 #region // Copyright (c) 2018, SIL International. All Rights Reserved.
-// <copyright from='2011' to='2018' company='SIL International'>
+// <copyright from='2018' to='2018' company='SIL International'>
 //		Copyright (c) 2018, SIL International. All Rights Reserved.
 //
 //		Distributable under the terms of the MIT License (http://sil.mit-license.org/)
@@ -11,7 +11,6 @@ using System;
 using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Windows.Forms;
-using HearThis.Properties;
 
 namespace HearThis.UI
 {
@@ -19,26 +18,22 @@ namespace HearThis.UI
 	{
 		private const int kHorizontalPadding = 4;
 		private const int kVerticalPadding = 4;
+		const TextFormatFlags kTextPositionFlags = TextFormatFlags.HorizontalCenter | TextFormatFlags.VerticalCenter;
 
 		private bool _selected;
 
 		private static int s_minWidth;
-		// The following should probably be overridden in derived classes, but
+		// The following property is intended to be overridden in derived classes, but
 		// this class is not abstract because it messes up Designer.
+		// ReSharper disable once UnassignedGetOnlyAutoProperty
 		protected virtual bool DisplayLabels { get; }
-		protected virtual Font LabelFont => Font;
 
-		protected static Font AttemptToCreateLabelFont(string fontFace, int size = 7)
+		protected virtual float LabelFontSize => 7;
+
+		public UnitNavigationButton()
 		{
-			try
-			{
-				return new Font(fontFace, size, FontStyle.Bold);
-			}
-			catch
-			{
-				// font face is probably missing.
-				return null;
-			}
+			if (Font.SizeInPoints != LabelFontSize)
+				Font = new Font(Font.FontFamily, LabelFontSize, FontStyle.Bold);
 		}
 
 		public bool Selected
@@ -54,7 +49,7 @@ namespace HearThis.UI
 			}
 		}
 
-		protected void DrawButton(Graphics g, int percentageTranslated, int percentageRecorded)
+		protected void DrawButton(Graphics g, bool hasTranslatedContent, int percentageRecorded)
 		{
 			if (Selected)
 				g.FillRectangle(AppPallette.HighlightBrush, 0, 0, Width, Height);
@@ -63,15 +58,18 @@ namespace HearThis.UI
 			int fillHeight = Height - kVerticalPadding;
 			var r = new Rectangle(kHorizontalPadding / 2, kVerticalPadding / 2, fillWidth, fillHeight);
 
-			Brush fillBrush = percentageTranslated > 0 ? AppPallette.BlueBrush : AppPallette.EmptyBoxBrush;
+			Brush fillBrush = hasTranslatedContent ? AppPallette.BlueBrush : AppPallette.EmptyBoxBrush;
 			g.FillRectangle(fillBrush, r);
 
 			g.SmoothingMode = SmoothingMode.AntiAlias;
 
-			if (Settings.Default.DisplayNavigationButtonLabels && percentageTranslated > 0 && LabelFont != null)
-				DrawLabel(g, r);
-			else
-				DrawProgressIndicators(g, r, percentageRecorded);
+			if (hasTranslatedContent)
+			{
+				if (DisplayLabels)
+					DrawLabel(g, r);
+				else if (percentageRecorded > 0)
+					DrawProgressIndicators(g, r, percentageRecorded);
+			}
 		}
 
 		protected void DrawProgressIndicators(Graphics g, Rectangle bounds, int percentageRecorded)
@@ -79,11 +77,12 @@ namespace HearThis.UI
 			// if it is selected, drawing this line just makes the selection box look irregular.
 			// Also, they can readily see what is translated in the selected book or chapter by
 			// looking at the more detailed display of its components.
-			if (percentageRecorded > 0 && percentageRecorded < 100 && !Selected)
+			if (percentageRecorded < 100)
 			{
-				g.DrawLine(AppPallette.CompleteProgressPen, bounds.Left, bounds.Bottom - 1, bounds.Right - 1, bounds.Bottom - 1);
+				if (!Selected)
+					g.DrawLine(AppPallette.CompleteProgressPen, bounds.Left, bounds.Bottom - 1, bounds.Right - 1, bounds.Bottom - 1);
 			}
-			else if (percentageRecorded >= 100)
+			else
 			{
 				int v1 = bounds.Height / 2 + 3;
 				int v2 = bounds.Height / 2 + 7;
@@ -95,13 +94,13 @@ namespace HearThis.UI
 					try
 					{
 						using (var font = new Font("Arial", 9, FontStyle.Bold))
-							TextRenderer.DrawText(g, "!", font, bounds, AppPallette.HilightColor,
-								TextFormatFlags.HorizontalCenter | TextFormatFlags.VerticalCenter);
+							DrawWarningIndicator(g, font, bounds);
 						return;
 					}
 					catch (Exception)
 					{
-						// Arial is probably missing. Just let it draw the check mark. Beats crashing.
+						// Arial is probably missing. Just use the default (label) font. Beats crashing.
+						DrawWarningIndicator(g, Font, bounds);
 					}
 				}
 
@@ -113,13 +112,16 @@ namespace HearThis.UI
 			}
 		}
 
+		private void DrawWarningIndicator(Graphics g, Font font, Rectangle bounds)
+		{
+			TextRenderer.DrawText(g, "!", font, bounds, AppPallette.HilightColor, kTextPositionFlags);
+		}
+
 		protected void DrawLabel(Graphics g, Rectangle bounds)
 		{
-			const TextFormatFlags positionFlags = TextFormatFlags.HorizontalCenter | TextFormatFlags.VerticalCenter;
 
 			bounds.Offset(0, -1);
-			TextRenderer.DrawText(g, Text, LabelFont, bounds,
-				DisplayLabels ? AppPallette.NavigationTextColor : AppPallette.MouseOverButtonBackColor, positionFlags);
+			TextRenderer.DrawText(g, Text, Font, bounds, AppPallette.NavigationTextColor, kTextPositionFlags);
 		}
 	}
 }